requirements.txt
Django>=5.1,<5.2
Pillow>=11.0.0,<12.0.0
django-imagekit>=5.0.0,<6.0.0
python-magic>=0.4.27,<0.5.0
django-cleanup>=9.0.0,<10.0.0
django-storages>=1.14.4,<1.15.0
pilkit>=3.0,<4.0
django-widget-tweaks>=1.5.0,<1.6.0
requests>=2.31.0,<3.0.0
